Контрольные вопросы [#219457] - Убрал Наследование.
